requirements.txt
instagrapi
google-generativeai
requests
python-dotenv
flask
pyrogram
tgcrypto
pymongo
